requirements/base.txt
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
python-dotenv==1.0.0
django-ckeditor==6.7.0
django-phonenumber-field==7.1.0
django-filter==23.2
django-cors-headers==4.2.0
phonenumbers==8.13.20
Pillow==10.0.0
